Bugfix: `if "false"` is always `True` (#617)
There is an issue in this launch file when passing the string 'false' as
an argument. In Python, non-empty strings are always evaluated as True,
regardless of their content. This means that even if you pass 'false',
the system will still evaluate it as True.

This bug results in the launch system incorrectly calling the OnShutdown
method twice. When any ROS launch action invokes a RosAdapter, it
triggers the following exception: "Cannot shutdown a ROS adapter that is
not running."

To temporarily work around this issue, you can launch gz_sim_launch.py
with the on_exit_shutdown argument set to an empty string. This prevents
the erroneous shutdown sequence and avoids the associated exception.

Signed-off-by: Ignacio Vizzo <ignacio@dexory.com>
(cherry picked from commit 1e30af0105058d68c8f1c98f37904505f613cf97)